requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
Brotli==1.0.9
cairocffi==1.5.1
CairoSVG==2.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.6
dj==0.1
Django==4.2.1
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.12.0
fonttools==4.39.4
freetype-py==2.3.0
html5lib==1.1
idna==3.4
lxml==4.9.2
mysqlclient==2.1.1
oscrypto==1.3.0
Pillow==9.5.0
pipenv==2023.4.29
pipreqs==0.4.13
platformdirs==3.5.1
pycairo==1.23.0
pycparser==2.21
pydyf==0.6.0
pyglet==2.0.7
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.9.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
rlPyCairo==0.2.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.2
virtualenv==20.23.0
virtualenv-clone==0.5.7
weasyprint==59.0
webencodings==0.5.1
xlwt==1.3.0
yarg==0.1.9
zopfli==0.2.2
